IFS-2182 - Stop InviteOrganisations from being deleted if they still contain Invites
This change also requires some additional changes to how the the
InviteOrganisation is persisted (we now `saveAndFlush` on the InviteOrganisation)
instead, so that the de-linking of the ApplicationInvite is cascaded down
first.

This stops a deletion of InviteOrganisation from deleting the Invite as
well.

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/invite/transactional/InviteServiceImpl.java
@@ -296,11 +296,15 @@ public class InviteServiceImpl extends BaseTransactionalService implements Invit
             InviteOrganisation inviteOrganisation
     ) {
         InviteOrganisation currentInviteOrganisation = invite.getInviteOrganisation();
+        currentInviteOrganisation.removeInvite(invite);
+        inviteOrganisationRepository.saveAndFlush(currentInviteOrganisation);
+
+        if (currentInviteOrganisation.getInvites().isEmpty()) {
+            inviteOrganisationRepository.delete(currentInviteOrganisation);
+        }
 
         invite.setInviteOrganisation(inviteOrganisation);
-        applicationInviteRepository.saveAndFlush(invite);
-
-        inviteOrganisationRepository.delete(currentInviteOrganisation);
+        applicationInviteRepository.save(invite);
     }
 
     private void initializeInvitee(ApplicationInvite invite, User user) {
